--- src/components/base-address/interfaces/address-interface.ts
@@ -1,9 +1,0 @@
-export interface AddressIF {
-  street: string
-  streetAdditional?: string
-  city: string
-  region: string
-  postalCode: string
-  country: string
-  deliveryInstructions?: string
-}
