--- src/components/base-address/interfaces/index.ts
@@ -1,2 +1,0 @@
-export { AddressIF } from './address-interface'
-export { SchemaIF } from './schema-interface'
